Moved SingleUsage class def, and updated comments
